+++ sydtest-yesod/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [0.2.0.0] - Unreleased
 
+### Added
+
+* Added `IsTest` instances for `YesodClientM site`
+
 ### Changed
 
 * Simplified `yit` using the new `ReaderT` instance in sydtest.